% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_gener_seqdata.R
\name{calc_EG_general}
\alias{calc_EG_general}
\title{Andriy's original get_EG function, more general case.}
\usage{
calc_EG_general(M, p, rdv)
}
\arguments{
\item{M:}{genotype likelihoods AA, Aa, aa, matrix sum(rdv) by 3 (double);
uses output from \code{calc_pobs_ndepth} function for simulation data and output from \code{getgenexp} or \code{getMAF} for VCF input}

\item{p:}{genotype frequencies AA, Aa, aa (double); output from \code{calc_EM} function.}

\item{rdv:}{read depth (vector of integers) for all samples, should be a vector of length=nsample.}
}
\value{
EG: a vector containing conditional expectation values. Size should be equal to length(rdv).
}
\description{
It is used in \code{generate_seqdata_null} and \code{generate_seqdata_alt}
 The more general case for \code{calc_EG}
}

