% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_gener_seqdata.R
\name{calc_EG_Var_general}
\alias{calc_EG_Var_general}
\title{Andriy's original cal_EG2}
\usage{
calc_EG_Var_general(M, p, rdv)
}
\arguments{
\item{M:}{genotype likelihoods AA, Aa, aa, matrix sum(rdv) by 3 (double);
#' uses output from \code{calc_pobs_ndepth} function for simulation data and output from \code{getgenexp} or \code{getMAF} for VCF input}

\item{p:}{genotype frequencies AA, Aa, aa (double); output from \code{calc_EM} function.}

\item{rdv:}{read depth (vector of integers) for all samples}
}
\value{
the variance of E(G_ij|D_ij)
}
\description{
Andriy's original cal_EG2
}
\section{Functions}{
\itemize{
\item \code{calc_EG_Var_general}: also, see \code{calc_EG_Var},
}}

