% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_gener_seqdata.R
\name{calc_EM}
\alias{calc_EM}
\title{EM calculation: Given a n by 3 matrix (M) containing likelihoods, this function uses the EM algorithm to estimate genotype frequencies.}
\usage{
calc_EM(M)
}
\arguments{
\item{M}{matrix of likelihoods derived from PL data.}
}
\value{
a vector of three values: p0, q0, 1-p0-q0
}
\description{
This part is the new part of the analysis where we can use reads instead of hard calls.
For this to work we need to estimate genotype frequencies.
Data consists of matrix n by 3
p = P(G=0)
q = P(G=1)
}

