% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_gener_seqdata.R
\name{calc_pobs_ndepth}
\alias{calc_Mr}
\alias{calc_pobs_ndepth}
\title{calc_pobs_ndepth:  Get likelihood P(Di|G) matrix given a vector of error rates and a vector of sequence reads.}
\usage{
calc_pobs_ndepth(Error, vect_row_reads, ndepth = 1)

calc_Mr(Error, vect_row_reads, ndepth = 1, rdv)
}
\arguments{
\item{Error:}{vector of base call error rates; Uses the same \code{error} input in the \code{seq_call} function
but concatenate all \code{error} values over all samples.}

\item{vect_row_reads:}{vector of row sequence reads ('A','G','T','C') outputted from function \code{seq_call},
but concatenated over all samples in same fashion as \code{Error}.}

\item{ndepth:}{integer, read depth, should be length(error)=length(vect_row_reads)=ndepth.}

\item{rdv:}{vector of read depths concatenated for all samples (integer).}

\item{Error:}{vector of base call error rates; Uses the same \code{error} input in the \code{seq_call} function
but concatenate all \code{error} values over all samples.}

\item{vect_row_reads:}{vector of row sequence reads ('A','G','T','C') outputted from function \code{seq_call},
but concatenated over all samples in same fashion as \code{Error}.}

\item{ndepth:}{integer, read depth, length(error)=length(vect_row_reads)=ndepth.}
}
\value{
M matrix of P(read|AA), P(read|Aa) and P(read|aa), where row dimensions equal \code{length(Error)}=length(vect_row_reads), by 3 columns.

M matrix of P(read|AA), P(read|Aa) and P(read|aa), where row dimensions equal \code{length(Error)}=length(vect_row_reads), by 3 columns.
Looks like getting likelihood over all reads cumulatively? Andriy, please explain.
}
\description{
This function calculates a likelihood matrix given a vector of error rates and a vector of sequence reads.
 It is called from the \code{generate_seqdata_null} and \code{generate_seqdata_alt} functions.
s  The output of this function  are used as inputs to \code{calc_EG}.
it use function \code{calc_psingle} to calculate the likelihood for each read and one possible genotype (TT, CT or CC)
Note to Andriy:  Why is LG hardcoded to LG=c('TT','CT','CC')  ??  Guess it's to denote the AA, Aa, aa.

It is used in \code{generate_seqdata_null} and \code{generate_seqdata_alt}
}
\section{Functions}{
\itemize{
\item \code{calc_Mr}: with additional param \code{rdv}
}}
\seealso{
\code{\link{seq_call}}
}

