% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_gener_seqdata.R
\name{calc_psingle}
\alias{calc_psingle}
\title{calc_psingle computes the genotype likelihood  P(D=g|G=G1G2,e) for a single base read given the sequencing error rate and true genotype.}
\usage{
calc_psingle(sing_read, A1, A2, error)
}
\arguments{
\item{sing_read:}{single read, ie. ('A')}

\item{A1:}{allele 1 in the true genotype, eg. (A1-'A',A2-'C')}

\item{A2:}{allele 2 in the true genotype, ie. (A1-'A',A2-'C')}

\item{error:}{sequencing error (double)}
}
\value{
p - likelihood of observing sing_read given genotype (A1A2) and error rate
}
\description{
calc_psingle is called from the \code{calc_pobs_ndepth} and \code{calc_Mr} functions.
please see Appendix B for how to calculate, P(D=g|G=G1G2)=1/2*P(D=g|G1)+1/2*P(D=g|G2) and  p(D=g|G1=g)=1-e, p(D=other letter|G1=g)=e/3.
}
\seealso{
\code{\link{calc_pobs_ndepth}} or \code{\link{calc_Mr}}
}

