% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_assoc_test.R
\name{calc_robust_Var}
\alias{calc_robust_Var}
\title{the robust variance estimation for case}
\usage{
calc_robust_Var(P)
}
\arguments{
\item{P:}{the genotype frequency, calcualted from EM algorithm}
}
\value{
the Robust variance of E(Gij|Dij)
called by \code{test_RVS_asy},\code{test_RVS_bootstrap}, and \code{calc_ScoreV_RVS}
}
\description{
Variance of genotypes Var(X)=E(X^2)-E(X)^2
}

