% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_assoc_test.R
\name{calc_ScoreV_RVS_regVar}
\alias{calc_ScoreV_RVS_regVar}
\title{calc_ScoreV_RVS_regVar: Get Variance of vector S by RVS using regular expected prob}
\usage{
calc_ScoreV_RVS_regVar(X1, X2, Y)
}
\arguments{
\item{X1:}{the expected conditional genotype probability for case}

\item{X2:}{the expected conditional genotype probability for control}

\item{Y:}{the phenotype}
}
\value{
the variance of score statistic vector S by RVS using regular expected prob
}
\description{
it is called in \code{test_rare_RVS_regVar}
}

