% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_seqdata.R
\name{generate_seqdata_alt}
\alias{generate_seqdata_alt}
\title{Generate sequence data under alternative hypothesis
it calls functions \code{seq_call},\code{calc_EG_general},\code{calc_EG_Var}, \code{calc_EM}, \code{calc_pobs_ndepth}
\code{calc_Mr}}
\usage{
generate_seqdata_alt(nsnp, ncase, ncont, mdcase, sddcase, mdcont, sddcont, em,
  esd, mmafCa, mmafCo)
}
\arguments{
\item{mmafCa}{MAF for nsnp variants in cases, HWE is assumed >0}

\item{mmafCo}{MAF for nsnp variants in controls, HWE is assumed >0}

\item{all}{the other params are describe in \code{generate_seqdata_null}, here MAF for cases/controls are different.}
}
\value{
For alt, outputs same object with additional:

MM - conditional expected value E(Gij|Dij)

P -estimated genotype frequencies P(G=0), P(G=1), P(G=2)

G - true genotype from which sequence data generated

MM2-  variance  of the conditional expected value E(Gij|Dij)
}
\description{
Generate sequence data under alternative hypothesis
it calls functions \code{seq_call},\code{calc_EG_general},\code{calc_EG_Var}, \code{calc_EM}, \code{calc_pobs_ndepth}
\code{calc_Mr}
}

