% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_seqdata.R
\name{generate_seqdata_null}
\alias{generate_seqdata_null}
\title{Generate sequence data under null hypothesis, it calls functions \code{seq_call},\code{calc_EG_general},\code{calc_EM},\code{calc_pobs_ndepth}
\code{calc_Mr}}
\usage{
generate_seqdata_null(nsnp, ncase, ncont, mdcase, sddcase, mdcont, sddcont, em,
  esd, mmaf)
}
\arguments{
\item{nsnp,}{number of variants (integer) >0}

\item{ncase,}{number of cases (integer) >0}

\item{ncont,}{number of controls (integer) >0}

\item{mdcase,}{average read depth in cases (double) >0}

\item{sddcase,}{standard deviation of read depth in cases  (double) >0}

\item{mdcont,}{average read depth in controls (double) >0}

\item{sddcont,}{standard deviation of read depth in controls (double) >0}

\item{em,}{average error rate, probability that sequence call is wrong (double) >0}

\item{esd,}{standard deviation of error rate (double) >0}

\item{mmaf,}{a vector of MAF for nsnp variants, HWE is assumed >0}
}
\value{
MM - conditional expected value E(Gij|Dij)

P -estimated genotype frequencies P(G=0), P(G=1), P(G=2)

G - true genotype from which sequence data generated
}
\description{
Generate sequence data under null hypothesis, it calls functions \code{seq_call},\code{calc_EG_general},\code{calc_EM},\code{calc_pobs_ndepth}
\code{calc_Mr}
}

