% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_vcf.R
\name{get_IDs}
\alias{get_IDs}
\title{Seprates the case and control IDs.}
\usage{
get_IDs(vcf_file, caseID_file, IDline = 128, headnum = 9)
}
\arguments{
\item{vcf_file,}{The 'address/filename' of the vcf file.}

\item{caseID_file,}{The 'address/filename' of the file including case IDs.}

\item{IDline,}{The number of the line in the vcf file whch contains the ID's, the provided vcf has IDline=128.}

\item{headnum,}{The number of columns in the headers before the IDs starts in the vcf file, default=9.}
}
\value{
a list which contains the position (column) numbers which correspond to the cases and controls in the row of the VCF with IDs, but columns are shifted by the headnum columns.
}
\description{
This function read the header of the vcf file (rows starting with # in vcf) and split the case and controls according to the caseID_file.
}
\examples{
getIDs("../data/1g113low_1g56exomehigh_filtered.hg19.chr11_1000snps.vcf","../data/bam.txt",128,9)
}

