% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_vcf.R
\name{get_likelihood_vcf}
\alias{get_likelihood_vcf}
\title{Gets Likelihoods for all possible genotypes given the set of alleles defined in the REF and ALT fields
This function fetches the genotype likelihoods from the VCF file (wo header) for the analysis and
must contain a 9th column (ie. 'FORMAT' field) consisting of (GT:AD:DP:GQ:PL).
The first step is to read the vcf file as a matrix and outputs an R object with 6 lists including 5 vectors and 1 list including 3 matrix:
 5 vectors: chr, position,ref allele, alt allele, filter indicator for 'PASS'
1 list: likelihood of L(00|ref,alt), L(01|ref,alt), L(11|ref,alt), which are three matrix with row for sample and col for variants}
\usage{
get_likelihood_vcf(M, nhead)
}
\arguments{
\item{M}{The vcf file without the first headers in a matrix (nsnps rows and nhead column +each column for one individuals).}

\item{nhead,}{the number of columns before the data for each individual}
}
\value{
Outputs an R object with:\cr
1. chromosome of variant
2. POS of variant location (vector)\cr
3. REF allele (vector)\cr
4. ALT allele (vector)\cr
5. FILTER indicator for 'PASS' (vector)\cr
6. likelihoods: L(D|0), L(D|1), L(D|2). List of 3 matrix.

L0 - matrix of P(D|0), rows are individuals and columns are SNPS\cr
L1 - matrix of P(D|1), rows are individuals and columns are SNPS\cr
L2 - matrix of P(D|2), rows are individuals and columns are SNPS\cr
}
\description{
Input matrix:
2nd column (POS)    - location
4th column (REF)    - reference allele
5th column (ALT)    - alternative allele
7th column (FILTER) - vcr filter value
10th column and onwards are 'data' columns in format (GT:AD:DP:GQ:PL)
}
\details{
GL = Phred-scaled likelihoods for genotypes as defined in the VCF specification format of 1000 Genome project,
likelihoods for all possible genotypes given the set of alleles defined in the REF and ALT fields.
Phred-scaled likelihoods: -log10(L(00|ref,alt)), -log10(L(01|ref,alt)), -log10(L(11|ref,alt)).
PL : the phred-scaled genotype likelihoods rounded to the closest integer (and otherwise defined precisely as the GL field) (Integers)

Example: 0/0:2,0:4:6:0,6,42
}

