% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_vcf.R
\name{get_lsingle_vcf}
\alias{get_lsingle_vcf}
\title{Parse Phred-scaled likelihoods}
\usage{
get_lsingle_vcf(Mij)
}
\arguments{
\item{M:}{a single data unit of format: GT:AD:DP:GQ:PL.  Example: get_l('0/0:2,0:4:6:0,6,42')}
}
\value{
a vector of three values: l0, l1, l2
}
\description{
This function parses the data column (GT:AD:DP:GQ:PL) to get the PL values \code{-10log10(x)},
and converts them to a vector of likelihoods: L0,L1,L2. It's called within \code{get_likelihood_vcf}.
}
\seealso{
\code{\link{get_likelihood_vcf}} for parent function.
}
\keyword{\code{get_likelihood_vcf}}
\keyword{by}
\keyword{called}

