% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_assoc_test.R
\name{rm_hom}
\alias{rm_hom}
\title{Input}
\usage{
rm_hom(case, cont, snp, type = "expected")
}
\arguments{
\item{case:}{a matrix for conditional expected genotype or genotype calls for case. Dimension ncase*nsnp}

\item{cont:}{a matrix for conditional expected genotype or genotype calls for controls. Dimension ncont*nsnp}

\item{snp:}{a matrix for snps, including chr, position, expected sample frequency if the case and controls are conditional expected genotype}
}
\value{
a vector indicate the snp columns that are monormophic.
}
\description{
Input
}

