% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tests.R
\name{RVS_asy}
\alias{RVS_asy}
\title{RVS using asymptotic distribution for score test statistic, the variance of score using regular formula for Var_case(E(G_ij|D_ij))
when method = Regular (also called likelihood method in the paper), Var_case(G_ij)  is used for Var_case(E(G_ij|D_ij)) when method = RVS.}
\usage{
RVS_asy(Y, X, P, method = "RVS")
}
\arguments{
\item{P}{(vector with length 3  (double)) - estimated genotype frequency for a variant P(G=0), P(G=1) and P(G=2)}

\item{method,}{has value RVS or regular to indicate the estimation of variance of the score statistic. using the population frequency P
to estimate variance for case if method=RVS, otherwise use the expected genotype probability for case.}

\item{Y:}{a vector with the phenotype for n individuals (=ncase+ncont), first ncase (y=1) then ncont (y=0)  (integar)}

\item{X:}{a vector of genotype for a snp, first ncase and then ncont}
}
\value{
p-values for the snp
}
\description{
Note, both function scaled the variance in equation (1) in Appendix A by dividing N_case*Ncontrol/N
also note the test statistics in anova gives us Rao=s^2/var(s^2), where RVS=s^2/robust var(s^2)
so RVS=Rao *var(s^2)/robust var(s^2), s=sum(y_j-mean(y))E(G_ij|D_ij), therefore var(s)//var(E(Gij|Dij)), var(X) in code.
}

