% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tests.R
\name{RVS_btrap}
\alias{RVS_btrap}
\title{use RVS to test associaton by bootrtrap, given phenotype (Y), expected values of genotypes for case and controls (X)
estimated genotype frequency (P) and number of times of bootstrap (nboot)}
\usage{
RVS_btrap(Y, X, P, nboot, method = "RVS")
}
\arguments{
\item{P}{(vector with length 3  (double)) - estimated genotype frequency for a variant P(G=0), P(G=1) and P(G=2)}

\item{method,}{has value RVS or regular to indicate the estimation of variance of the score statistic. using the population frequency P
to estimate variance for case if method=RVS, otherwise use the expected genotype probability for case.}

\item{Y:}{a vector with the phenotype for n individuals (=ncase+ncont), first ncase (y=1) then ncont (y=0)  (integar)}

\item{X:}{a vector of genotype for a snp, first ncase and then ncont}

\item{nboot:}{number of bootstrap}
}
\value{
p-values for the snp
}
\description{
use RVS to test associaton by bootrtrap, given phenotype (Y), expected values of genotypes for case and controls (X)
estimated genotype frequency (P) and number of times of bootstrap (nboot)
}

