% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_assoc_test.R
\name{sample_bootstrap}
\alias{sample_bootstrap}
\title{helper11: Bootstrap resampling}
\usage{
sample_bootstrap(X1, X2)
}
\arguments{
\item{X1:}{genotype likelihood matrix for case}

\item{X2:}{genotype likelihood matrix for control}
}
\value{
X: bootstrap for genotype likelihood matrix for case and control
}
\description{
it calls \code{calc_centralized_matrix} and called by \code{\link{test_rare_RVS}}, \code{\link{test_rare_RVS_regVar}}.
}

