% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_gener_seqdata.R
\name{seq_call}
\alias{seq_call}
\title{seq_call produces the base call from given a genotype and the error scores at one locus.}
\usage{
seq_call(genotype, error, ndepth)
}
\arguments{
\item{genotype:}{vector of two alleles for a given genotype, ie. c('C', 'T')}

\item{error:}{vector of base call errors: Phen scores (double).  Length of this vector is the read depth.}

\item{ndepth:}{the read depth (integer)}
}
\value{
vect_reads: vector of row sequence reads ('A','G','T','C') for single variant. The length of this vector is the read depth.
}
\description{
It is called from the \code{generate_seqdata_null} and \code{generate_seqdata_alt} functions.
}

