% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{snp_true_2T}
\alias{snp_true_2T}
\title{snp_true_2T is used to change all TRUE genotype into T (if all the individual has genotype 'T' in one strand, R automatically change them into 'TRUE')}
\usage{
snp_true_2T(M)
}
\arguments{
\item{M,}{the genotype matrix for one sample}
}
\value{
M, the genotype matrix after change TRUE into T
}
\description{
snp_true_2T is used to change all TRUE genotype into T (if all the individual has genotype 'T' in one strand, R automatically change them into 'TRUE')
}

