% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_assoc_test.R
\name{test_CAST}
\alias{test_CAST}
\title{helper7: the linear test for nsnp rare variant}
\usage{
test_CAST(S, Sigma, w, nsnp)
}
\arguments{
\item{S:}{vector of length nsnp, the Score for each variant}

\item{Sigma:}{matrix Sigma is the variance matrix of vector S.}

\item{w:}{the weight for CAST test}

\item{nsnp:}{number of snp}
}
\value{
p-values from CAST test for J variants (double)
}
\description{
it is called in \code{test_rare_regScore}, \code{test_rare_RVS} and \code{test_rare_RVS_regVar}
}

