% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_assoc_test.R
\name{test_Hotelling}
\alias{test_Hotelling}
\title{helper9:  Hotelling test
Input:'}
\usage{
test_Hotelling(S, Sigma)
}
\arguments{
\item{S:}{vector of length nsnp, the Score for each variant}

\item{Sigma:}{matrix Sigma is the variance matrix of vector S.}

\item{nsnp:}{number of snp}
}
\value{
p-values from Hotelling test for J variants (double)
}
\description{
helper9:  Hotelling test
Input:'
}

